Revert most of last commit
The method `isAllowedAny()` can't be used at this point of execution, because
it calls the `getSession()` method of the current request, which in turn
calls the `provideSessionInfo()` of us again. This ends up in an endless loop
and a stack overflow for new users who don't have an account yet.

Change-Id: I0a374120ba1ede78f9a7258914f738a9a40eaaa3